Work on supporting the first videocapture scenario
This adds a throwaway test that starts recording, begins an animation,
ends the animation, stops recording, and shuts down, along with various
infrastructure to support it. Specifically, we make some changes to
ffprocess_remote.py makes running the fennec binary non-blocking, and add
a bunch of code to control the firefox process to bcontroller.py.

This is just prototype code and needs tons of refinement and thought. It
probably breaks talos for the non-videocapture cases.